+++ nodes/Freedom24/methods/index.ts
@@ -1,9 +1,8 @@
-import { freedom24ApiCredentialTest, freedom24UserApiCredentialTest } from './credentialTest';
+import { freedom24UserApiCredentialTest } from './credentialTest';
 import { searchTickers } from './listSearch';
 
 export const methods = {
 	credentialTest: {
-		freedom24ApiCredentialTest,
 		freedom24UserApiCredentialTest,
 	},
 	listSearch: {